Добавил фабрику MessageDaoSingleton. Написал класс MessageDao!

+++ src/main/java/by/it_academy/jd2/Mk_JD2_95_22/dao/api/IMessageDao.java
@@ -8,5 +8,5 @@ public interface IMessageDao {
     List<MessageDTO> getMessage(String toUser);
     boolean setMessage(MessageDTO messageDTO);
 
-    int countMassage();
+    long countMassage();
 }
